requirements.txt
streamlit>=1.40.0
openai>=1.40.0
pyyaml>=6.0.1
python-docx
pandas
openpyxl
requests
reportlab
pillow
